Check if font is already installed

+++ install-requirements.sh
@@ -10,5 +10,7 @@ apt install git curl gcc make clangd python3 python3-venv python-is-python3 pip 
 # python3-pip install --user --upgrade pynvim ## replaced with apt install python3-nvim
 
 # Install JetBrainsMono Nerd Font
-curl -s -L -o JerBrainsMono.tar.xz https://github.com/ryanoasis/nerd-fonts/releases/download/v3.2.1/JetBrainsMono.tar.xz
-tar -xf JetBrainsMono.tar.xz -C /usr/local/share/fonts
+if [[ ! -e /usr/local/share/fonts/JetBrainsMonoNerdFont-Regular.ttf ]]; then
+  curl -s -L -o JerBrainsMono.tar.xz https://github.com/ryanoasis/nerd-fonts/releases/download/v3.2.1/JetBrainsMono.tar.xz
+  tar -xf JetBrainsMono.tar.xz -C /usr/local/share/fonts
+fi
